Redo redux setup
