+++ src/pages/sphere-page.js
@@ -15,7 +15,7 @@ const SpherePage = ({data}) => <Sphere data={data} />;
 // This is the magic provided by react-redux, you can pass props to the component like Scotty beams up Kirk!
 const mapStateToProps = appState => {
   return {
-    data: appState.sphere.data
+    data: appState.components.commits.data
   };
 };
