--- src/store/reducers/code-crawl.js
@@ -1,8 +1,0 @@
-// import C from '../../constants';
-import initialState from "../initial-state";
-
-export default (currentState = initialState.codeCrawl, action) => {
-  switch (action.type) {
-  default: return currentState;
-  }
-};
