+++ src/store/reducers/components.js
@@ -1,6 +1,7 @@
+// import C from '../../constants';
 import initialState from "../initial-state";
 
-export default (currentState = initialState.matrix, action) => {
+export default (currentState = initialState.components, action) => {
   switch (action.type) {
   default: return currentState;
   }
