--- src/store/reducers/earth.js
@@ -1,8 +1,0 @@
-// import C from '../../constants';
-import initialState from "../initial-state";
-
-export default (currentState = initialState.sphere, action) => {
-  switch (action.type) {
-  default: return currentState;
-  }
-};
